cli/pyproject.toml
[project]
name = "specsmith"
version = "0.1.0"
description = "Spec-driven development from the terminal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "SpecSmith Contributors" }]
keywords = ["spec", "development", "cli", "planning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
]
dependencies = [
    "typer[all]>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ai = ["anthropic>=0.40"]

[project.scripts]
specsmith = "specsmith.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specsmith"]

[tool.pytest.ini_options]
testpaths = ["tests"]
